requirements.txt
absl-py==1.0.0
attrs==21.4.0
cycler==0.11.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pygame==2.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
